refactor: remove SPL instructions + use message_id string type
- Ported from
https://github.com/axelarnetwork/axelar-amplifier-solana/commit/2da7e82ef567cfbdb1704fab69080e071dc4c646
- Removes all SPL related instructions but changes the events to be
future-compatible with SPL tokens. - Changes the add_gas and refund_fees
to accept a message_id of type String. - Changes the instruction names
to match Stellar's implementation. (ie. removes the _for_contract_call
in pay_gas). - Removes the treasury PDA in events as this is
determinstic and could be derived from the gas service program id. -
Adds sender to certain events to match Stellar's events.